Data and Source v1.13 Mod High Resolution version from 2006-04-15
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@25 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Laptop/BobbyRGuns.cpp
@@ -1209,8 +1209,8 @@ void SetFirstLastPagesForNew( UINT32 uiClassMask )
 	UINT16 i;
 	INT16	sFirst = -1;
 	INT16	sLast = -1;
-	UINT8	ubPages=0;
-	UINT8	ubNumItems=0;
+	UINT16	ubPages=0;
+	UINT16	ubNumItems=0;
 
 	gubCurPage = 0;
 
@@ -1254,8 +1254,8 @@ void SetFirstLastPagesForUsed()
 	UINT16 i;
 	INT16	sFirst = -1;
 	INT16	sLast = -1;
-	UINT8	ubPages=0;
-	UINT8	ubNumItems=0;
+	UINT16	ubPages=0;
+	UINT16	ubNumItems=0;
 
 	gubCurPage = 0;
