+++ Tactical/Rain.cpp
@@ -198,7 +198,7 @@ void ResetRain()
 
 void GenerateRainDropsList()
 {
-	guiCurrMaxAmountOfRainDrops = (UINT32)BASE_MAXIMUM_DROPS * gbCurrentRainIntensity;
+	guiCurrMaxAmountOfRainDrops = (UINT32)(BASE_MAXIMUM_DROPS * gbCurrentRainIntensity);
 
 	pRainDrops = (TRainDrop *)MemAlloc( sizeof( TRainDrop ) * guiCurrMaxAmountOfRainDrops );
 	memset( pRainDrops, 0, sizeof( TRainDrop ) * guiCurrMaxAmountOfRainDrops );
@@ -235,8 +235,8 @@ void CreateRainDrops()
 	BOOLEAN fLoopIsDone;
 	UINT32 uiIndRand;
 
-	fpCos = cos( DEGREE(fpCurrDropAngleOfFalling) );
-	fpSin = sin( DEGREE(fpCurrDropAngleOfFalling) );
+	fpCos = (FLOAT)(cos( DEGREE(fpCurrDropAngleOfFalling) ));
+	fpSin = (FLOAT)(sin( DEGREE(fpCurrDropAngleOfFalling) ));
 
 	if( fpCos )
 		fpAbsTg = fabs( fpSin / fpCos ); // take only absolute valuse
@@ -288,20 +288,20 @@ void CreateRainDrops()
 
 			if( uiIndex >= fpNumDropsToXBorder )
 			{
-				pCurr->fpX = gRainRegion.left + Random( gRainRegion.right - gRainRegion.left );
-				pCurr->fpY = gRainRegion.bottom - 1;
+				pCurr->fpX = (FLOAT)(gRainRegion.left + Random( gRainRegion.right - gRainRegion.left ));
+				pCurr->fpY = (FLOAT)(gRainRegion.bottom - 1);
 			}
 			else if( fpCos > 0 )
 			{
-				pCurr->fpX = gRainRegion.right - 1;
-				pCurr->fpY = gRainRegion.top + Random( gRainRegion.bottom - gRainRegion.top );
+				pCurr->fpX = (FLOAT)(gRainRegion.right - 1);
+				pCurr->fpY = (FLOAT)(gRainRegion.top + Random( gRainRegion.bottom - gRainRegion.top ));
 			}else{
 				pCurr->fpX = 0;
-				pCurr->fpY = gRainRegion.top + Random( gRainRegion.bottom - gRainRegion.top );
+				pCurr->fpY = (FLOAT)(gRainRegion.top + Random( gRainRegion.bottom - gRainRegion.top ));
 			}
 		}else{
-			pCurr->fpX = gRainRegion.left + Random( gRainRegion.right - gRainRegion.left );
-			pCurr->fpY = gRainRegion.top + Random( gRainRegion.bottom - gRainRegion.top );
+			pCurr->fpX = (FLOAT)(gRainRegion.left + Random( gRainRegion.right - gRainRegion.left ));
+			pCurr->fpY = (FLOAT)(gRainRegion.top + Random( gRainRegion.bottom - gRainRegion.top ));
 		}
 
 		pCurr->uiAmountOfTicksToLive = 0;
@@ -405,7 +405,7 @@ void RenderRainOnSurface()
 
 		if( !pCurr->fAlive )continue;
 		
-		LineDraw( TRUE, pCurr->fpX, pCurr->fpY, pCurr->fpX + pCurr->fpEndRelX, pCurr->fpY + pCurr->fpEndRelY,  sDropsColor, pDestBuf );
+		LineDraw( TRUE, (int)pCurr->fpX, (int)pCurr->fpY, (int)pCurr->fpX + (int)pCurr->fpEndRelX, (int)(pCurr->fpY + pCurr->fpEndRelY),  sDropsColor, pDestBuf );
 	}
 
 	UnLockVideoSurface( guiRainRenderSurface );
@@ -427,35 +427,35 @@ void GenerateRainMaximums()
 		fpMaxDropAngleOfFalling = 160;
 	}
 
-	fpCurrDropAngleOfFalling = fpMinDropAngleOfFalling + Random( fpMaxDropAngleOfFalling - fpMinDropAngleOfFalling );
+	fpCurrDropAngleOfFalling = fpMinDropAngleOfFalling + Random( (UINT32)(fpMaxDropAngleOfFalling - fpMinDropAngleOfFalling) );
 	
 	fpMinDropLength = MIN_DROP_LENGTH + ADD_DROP_LENGTH_IF_STORM * ( gbCurrentRainIntensity - 1 );
 	fpMaxDropLength = fpMinDropLength + DROP_LENGTH_RANGE;
 
-	fpCurrDropLength = fpMinDropLength + Random( fpMaxDropLength - fpMinDropLength );
+	fpCurrDropLength = fpMinDropLength + Random( (UINT32)(fpMaxDropLength - fpMinDropLength) );
 
 	fpMinDropSpeed = BASE_DROP_SPEED * gbCurrentRainIntensity;
 	fpMaxDropSpeed = fpMinDropSpeed + DROP_SPEED_RANGE;
 
-	fpCurrDropSpeed = fpMinDropSpeed + Random( fpMaxDropSpeed - fpMinDropSpeed );
+	fpCurrDropSpeed = fpMinDropSpeed + Random( (UINT32)(fpMaxDropSpeed - fpMinDropSpeed) );
 
 }
 
 void UpdateRainDropsProperities()
 {
-	fpCurrDropAngleOfFalling += Random( 1000 * DROP_ANGLE_CHANGE_RATE * gbCurrentRainIntensity * 2 ) / 1000.0f - DROP_ANGLE_CHANGE_RATE * gbCurrentRainIntensity;
+	fpCurrDropAngleOfFalling += Random( (UINT32)(1000 * DROP_ANGLE_CHANGE_RATE * gbCurrentRainIntensity * 2 )) / 1000.0f - DROP_ANGLE_CHANGE_RATE * gbCurrentRainIntensity;
 
 	fpCurrDropAngleOfFalling = max( fpMinDropAngleOfFalling, fpCurrDropAngleOfFalling );
 	fpCurrDropAngleOfFalling = min( fpMaxDropAngleOfFalling, fpCurrDropAngleOfFalling );
 
 
-	fpCurrDropLength += Random( 1000 * DROP_LENGTH_CHANGE_RATE * 2 ) / 1000.0f - DROP_LENGTH_CHANGE_RATE;
+	fpCurrDropLength += Random( (UINT32)(1000 * DROP_LENGTH_CHANGE_RATE * 2 )) / 1000.0f - DROP_LENGTH_CHANGE_RATE;
 
 	fpCurrDropLength = max( fpMinDropLength, fpCurrDropLength );
 	fpCurrDropLength = min( fpMaxDropLength, fpCurrDropLength );
 
 
-	fpCurrDropSpeed += Random( 1000 * DROP_SPEED_CHANGE_RATE * 2 ) / 1000.0f - DROP_SPEED_CHANGE_RATE;
+	fpCurrDropSpeed += Random( (UINT32)(1000 * DROP_SPEED_CHANGE_RATE * 2 )) / 1000.0f - DROP_SPEED_CHANGE_RATE;
 
 	fpCurrDropSpeed = max( fpMinDropSpeed, fpCurrDropSpeed );
 	fpCurrDropSpeed = min( fpMaxDropSpeed, fpCurrDropSpeed );
